backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
google-cloud-documentai>=3.0.0
google-genai>=1.0.0
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.0.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
pydantic>=2.5.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
